@@ -9,7 +9,7 @@ import (
 func main() {
 	app := pine.New()
 
-	di.Set("render", func(builder di.AbstractBuilder) (i interface{}, e error) {
+	di.Set("render", func(builder di.AbstractBuilder) (i any, e error) {
 		// reload = true 每次都会重载模板
 		return ptemplate.New("views", ".html", false), nil
 	}, true)
