@@ -8,21 +8,21 @@ import (
 	"errors"
 )
 
-type RememberCallback func() (interface{}, error)
+type RememberCallback func() (any, error)
 
 type AbstractCache interface {
 	Get(string) ([]byte, error)
-	GetWithUnmarshal(string, interface{}) error
+	GetWithUnmarshal(string, any) error
 
 	Set(string, []byte, ...int) error
-	SetWithMarshal(string, interface{}, ...int) error
+	SetWithMarshal(string, any, ...int) error
 
 	Delete(string) error
 	Exists(string) bool
 
-	Remember(string, interface{}, RememberCallback, ...int) error
+	Remember(string, any, RememberCallback, ...int) error
 
-	GetProvider() interface{}
+	GetProvider() any
 }
 
 var ErrKeyNotFound = errors.New("key not found or expired")
