@@ -28,7 +28,7 @@ func (r *pLeveldb) Get(key string) (byts []byte, err error) {
 	return
 }
 
-func (r *pLeveldb) GetWithUnmarshal(key string, receiver interface{}) (err error) {
+func (r *pLeveldb) GetWithUnmarshal(key string, receiver any) (err error) {
 	var byts []byte
 	if byts, err = r.Get(key); err == nil {
 		err = cache.UnMarshal(byts, receiver)
@@ -43,7 +43,7 @@ func (r *pLeveldb) Set(key string, val []byte, ttl ...int) (err error) {
 	return err
 }
 
-func (r *pLeveldb) SetWithMarshal(key string, data interface{}, ttl ...int) error {
+func (r *pLeveldb) SetWithMarshal(key string, data any, ttl ...int) error {
 	if byts, err := cache.Marshal(data); err != nil {
 		return err
 	} else {
@@ -55,14 +55,14 @@ func (r *pLeveldb) Delete(key string) error {
 	return r.DB.Delete([]byte(key), &opt.WriteOptions{Sync: true})
 }
 
-func (r *pLeveldb) Remember(key string, receiver interface{}, call cache.RememberCallback, ttl ...int) (err error) {
+func (r *pLeveldb) Remember(key string, receiver any, call cache.RememberCallback, ttl ...int) (err error) {
 	defer func() {
 		if recoverErr := recover(); recoverErr != nil {
 			err = recoverErr.(error)
 		}
 	}()
 	if err = r.GetWithUnmarshal(key, receiver); cache.IsErrKeyNotFound(err) {
-		var value interface{}
+		var value any
 		if value, err = call(); err == nil {
 			if err = r.SetWithMarshal(key, receiver, ttl...); err == nil {
 				reflect.ValueOf(receiver).Elem().Set(reflect.ValueOf(value).Elem())
@@ -72,7 +72,7 @@ func (r *pLeveldb) Remember(key string, receiver interface{}, call cache.Remembe
 	return
 }
 
-func (r *pLeveldb) GetProvider() interface{} { return r.DB }
+func (r *pLeveldb) GetProvider() any { return r.DB }
 
 func (r *pLeveldb) Exists(key string) bool {
 	var err error
